+++ include/drivers/nvme.h
@@ -336,7 +336,7 @@ int nvme_controller_count(void);
 nvme_controller_t *nvme_get_controller(int i);
 
 /* Backend I/O entry points — called through blockdev ops table */
-int nvme_read_sectors(const struct blockdev_device *dev, uint32_t lba, uint32_t count, void *buffer);
-int nvme_write_sectors(const struct blockdev_device *dev, uint32_t lba, uint32_t count, const void *buffer);
+int nvme_read_sectors(const struct blockdev_device *dev, uint64_t lba, uint32_t count, void *buffer);
+int nvme_write_sectors(const struct blockdev_device *dev, uint64_t lba, uint32_t count, const void *buffer);
 
 #endif /* INCLUDE_DRIVERS_NVME_H_ */
